Add files via upload
Some fixes for the scouts analysis

This is now completely independent from anyone submitting their scouts the previous week with the scouting tool I made and doesn't rely on those scouted teams in the excel file matching the team in the MDS

It should just find the scouted pages in your MDS and print those out in jupyter. I've tested it on about 5 or 6 of my turns and it seems to work, though I can't guarantee it's bomb proof!